+++ common/src/traits/mod.rs
@@ -1,1 +1,1 @@
-
+pub mod query;
