+++ common/src/types/mod.rs
@@ -2,3 +2,8 @@ pub mod api;
 pub mod query;
 pub mod smt;
 pub mod tx_builder;
+
+pub use ethereum_types::{
+    BigEndianHash, Bloom, Public, Secret, Signature, H128, H160, H256, H512, H520, H64, U128,
+    U256, U512, U64,
+};
